+++ acmed/src/acme_proto/account.rs
@@ -1,5 +1,5 @@
 use crate::acme_proto::http;
-use crate::acme_proto::structs::{Account, AccountResponse, Directory};
+use crate::acme_proto::structs::Account;
 use crate::certificate::Certificate;
 use crate::endpoint::Endpoint;
 use crate::jws::algorithms::SignatureAlgorithm;
@@ -17,32 +17,25 @@ pub struct AccountManager {
 
 impl AccountManager {
     pub fn new(
-        cert: &Certificate,
-        directory: &Directory,
-        endpoint: &Endpoint,
-        nonce: &str,
+        endpoint: &mut Endpoint,
         root_certs: &[String],
-    ) -> Result<(Self, String), Error> {
+        cert: &Certificate,
+    ) -> Result<Self, Error> {
         // TODO: store the key id (account url)
         let key_pair = storage::get_account_keypair(cert)?;
+        let kp_ref = &key_pair;
         let account = Account::new(cert, endpoint);
         let account = serde_json::to_string(&account)?;
-        let data_builder =
-            |n: &str| encode_jwk(&key_pair, account.as_bytes(), &directory.new_account, n);
-        let (acc_rep, account_url, nonce): (AccountResponse, String, String) = http::get_obj_loc(
-            cert,
-            root_certs,
-            &directory.new_account,
-            &data_builder,
-            &nonce,
-        )?;
+        let acc_ref = &account;
+        let data_builder = |n: &str, url: &str| encode_jwk(kp_ref, acc_ref.as_bytes(), url, n);
+        let (acc_rep, account_url) = http::new_account(endpoint, root_certs, &data_builder)?;
         let ac = AccountManager {
             key_pair,
             account_url,
             orders_url: acc_rep.orders.unwrap_or_default(),
         };
         // TODO: check account data and, if different from config, update them
-        Ok((ac, nonce))
+        Ok(ac)
     }
 }