+++ CHANGELOG.md
@@ -13,6 +13,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Changed
+- The HTTP(S) part is now handled by `reqwest` instead of `http_req`.
+
+
 ## [0.7.0] - 2020-03-12
 
 ### Added